requirements.txt
annotated-types==0.6.0
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
et-xmlfile==1.1.0
filelock==3.12.4
greenlet==3.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
parsel==1.8.1
pdf2image==1.16.3
Pillow==10.0.1
pluggy==1.3.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
Pygments==2.16.1
pyOpenSSL==23.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.2
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
rich==13.6.0
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.21
tldextract==5.0.0
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
urllib3==2.0.6
w3lib==2.1.2
zope.interface==6.1
